% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{allclose}
\alias{allclose}
\title{Check whether all elements of array or vector like objects are the same}
\usage{
allclose(x, y, tol = 1e-12)
}
\arguments{
\item{x}{First array like object}

\item{y}{Second array like object}

\item{tol}{Acceptable difference between individual elements of \code{x}
and \code{y}}
}
\value{
logical indicating if values are elementwise within \code{tol}
of each other
}
\description{
i.e. within a small but non-zero tolerance of each other
}
