% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot.cv.glmnet}
\alias{autoplot.cv.glmnet}
\title{Plot cross-validated elastic net paths}
\usage{
\method{autoplot}{cv.glmnet}(object, rule = "1-se")
}
\arguments{
\item{object}{Object of class \code{cv.glmnet}}

\item{rule}{Either "min" or "1-se". Specifies whether lambda should be
selected to minimize CV error, or to find the sparsest model within one standard
deviation of the minimum error. Default \code{cv.glmnet} behavior is to use the
one standard error rule, following Breiman.}
}
\value{
ggplot2 object with elastic net paths
}
\description{
Plots paths for features in the final model in color and provides a legend
for these features only. Features not in the final are plotted in grey and not
included in the legend.
}
\examples{

library(glmnet)
data(Boston, package = "MASS")

X <- model.matrix(medv ~ ., Boston)
y <- Boston$medv
cv_model <- cv.glmnet(X, y)

autoplot(cv_model, rule = "min")

}
