% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{beta_ols}
\alias{beta_ols}
\title{Get OLS regression coefficients}
\usage{
beta_ols(X, y)
}
\arguments{
\item{X}{Data matrix (\code{n} by \code{p})}

\item{y}{Response vector (length \code{n})}
}
\value{
Vector of regression coefficients
}
\description{
Assumes data matrix is tall or square, but not wide. Does not add
  intercept.
}
