% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{combine}
\alias{combine}
\title{Semi-general \code{unite} to vectorize a function across columns of dataframe}
\usage{
combine(df, col, ..., .f, remove = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{col}{Bare/unquoted name of results column}

\item{...}{Bare/unquoted names of argument columns}

\item{.f}{A function that accepts as many arguments as provided argument
columns. Gets passed to \code{purrr::pmap_chr} so formula-style lambda
specification also works.}

\item{remove}{Whether or not to remove the argument columns (defaults
to \code{true})}
}
\value{
Dataframe with new column generated by applying \code{fun} to
argument columns element-wise
}
\description{
Accepts columns from a dataframe and vectorizes/parallel maps a function
across them, returning the result in a new column. Function must return a
character vector because \code{purrr::pmap_char} enforces type-safety.
}
