% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{compare_coefs}
\alias{compare_coefs}
\title{Compare coefficients values across two different models}
\usage{
compare_coefs(model1, model2, ...)
}
\arguments{
\item{model1}{A model object (from \code{lm} for example) with a \code{tidy} method.}

\item{model2}{A model object (from \code{lm} for example) with a \code{tidy} method.}

\item{...}{Bare (unquoted) coefficient names present in both models.}
}
\value{
A tidy tibble with:
\itemize{
\item the coefficient names
\item the difference in coefficient values between models
\item standard error of the difference
\item z-statistic for the difference
\item p-value of the difference
\item lower 95 percent confidence bound for the difference
\item upper 95 percent confidence bound for the difference
\item the coefficient estimate from model1
\item the coefficient estimate from model2
\item the comparison method, always "Z-test"
}
}
\description{
Compare coefficients values across two different models
}
\examples{

model <- lm(mpg ~ hp + drat, mtcars)
other <- lm(mpg ~ hp * disp + drat, mtcars)

compare_coefs(model, other, hp, drat)

}
