% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{constants_like}
\alias{constants_like}
\title{Create object of same shape and type as \code{x} filled with value \code{c}}
\usage{
constants_like(x, c)
}
\arguments{
\item{x}{Object to imitiate}

\item{c}{Value to fill imitation with}
}
\value{
Object of same shape and type as \code{x} filled with value \code{c}
}
\description{
Again copying the Numpy zeros_like and ones_like interface
}
