% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{count_non_na}
\alias{count_non_na}
\title{Count non-NA values in a vector}
\usage{
count_non_na(x)
}
\arguments{
\item{x}{vector}
}
\value{
number of non-NA values in the vector, 0 if there are none
}
\description{
Count non-NA values in a vector
}
