% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{diag_like}
\alias{diag_like}
\title{Create an identity matrix matching a vector}
\usage{
diag_like(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An identity matrix with the same dimension as x.
}
\description{
Create an identity matrix matching a vector
}
\examples{

diag_like(1:2)
}
