% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvn_arma}
\alias{dmvn_arma}
\title{Calculate multivariate normal densities}
\usage{
dmvn_arma(x, mean, sigma, log = FALSE)
}
\arguments{
\item{x}{Matrix of multivariate normal samples}

\item{mean}{Center of the distribution.}

\item{sigma}{Covariance matrix.}

\item{log}{Logical indicating if log density should be used.}
}
\value{
Matrix with single column of densities.
}
\description{
Calculate multivariate normal densities
}
\examples{

mu <- 1:3
sigma <- diag(3)
samps <- rmvn(10, mu, sigma)
dmvn(samps, mu + 1, sigma)

}
