% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_shareable_link_to_data}
\alias{get_shareable_link_to_data}
\title{Get an immediately shareable / pasteable link to data}
\usage{
get_shareable_link_to_data(data, path, direct = TRUE)
}
\arguments{
\item{data}{A data frame, tibble or other object that be written to
a CSV.}

\item{path}{What to call the data file, as a character. Can also be a
file path for more precise placement of the data file. Path must
exist both locally and in Google Drive (really, I'm not sure what
happens when it doesn't).}

\item{direct}{If \code{TRUE}, the link can be used immediately in R to
read in the file. If \code{FALSE}, the link is not directly to the CSV,
but to the Google Drive preview of the file, which allows for nice
previewing of the data.}
}
\value{
Invisibly returns the link as a character vector.
}
\description{
Uploads a data frame to Google Drive, make it shareable, and
copies the shareable link onto the clipboard. Requires that OAuth
has been set up. \code{googledrive} should help you through this if you
haven't set OAuth up already.
}
\details{
Creates an intermediate local CSV that is deleted after upload is
complete.
}
\examples{

# link for direct download / immediate use
lnk <- get_shareable_link_to_data(mtcars, "mtcars.csv")
mt <- read.csv(lk)
head(mt)

# link to nice preview
get_shareable_link_to_data(mtcars, "mtcars.csv", direct = FALSE)
}
