% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggpie}
\alias{ggpie}
\title{Create a pie chart in ggplot2}
\usage{
ggpie(df, main, labels = NULL, condition = NULL)
}
\arguments{
\item{df}{data frame}

\item{main}{column name of response variable}

\item{labels}{column name with response labels}

\item{condition}{column name of grouping label}
}
\value{
ggplot2 object
}
\description{
Please only do this if forced to do so by a boss, etc.
}
