% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayeslib-package.r
\docType{package}
\name{hayeslib}
\alias{hayeslib}
\alias{hayeslib-package}
\title{hayeslib: a package to learn about packages. #meta}
\description{
This packages provides functions in R/ and data analyses in vignettes/
}
\section{hayeslib functions}{

oh look this is how a section works
}

