% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.r
\name{indices}
\alias{indices}
\title{Returns the sequence of indices for a vector or list object.}
\usage{
indices(x)
}
\arguments{
\item{x}{Vector or list}
}
\value{
Indices for vector or list input. Otherwise throws an error.
}
\description{
Returns the sequence of indices for a vector or list object.
}
\examples{
indices(c("mouse", "cat", "dog"))

}
