% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{linspace}
\alias{linspace}
\title{Returns evenly spaced numbers.}
\usage{
linspace(start, stop, num = 15)
}
\arguments{
\item{start}{base^start is the starting value of the sequence}

\item{stop}{base^stop is the final value in the sequence}

\item{num}{Number of sequence elements to generate. Default is 15.}
}
\value{
Linearly spaced sequence from \code{start} to \code{stop}
}
\description{
This wraps seq trivially. What can I say, I like the Numpy API.
}
\examples{
linspace(-5, 5)

}
