% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lzy_chr}
\alias{lzy_chr}
\title{Create character vector without quotes}
\usage{
lzy_chr(...)
}
\arguments{
\item{...}{Bared, comma separated strings}
}
\value{
The bare code you wrote, but comma separated
}
\description{
Exactly the same as Hmisc::Cs except ain't nobody got time to import that.
}
\examples{
lzy_chr(this, will, become, a, character, vector)

}
