% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{lzy_impute}
\alias{lzy_impute}
\title{Mean impute and optionally add missingness indicator}
\usage{
lzy_impute(df, ..., indicate = TRUE)
}
\arguments{
\item{df}{data frame}

\item{...}{variable selections as per tidyselect}

\item{indicate}{logical, whether or not to add missingness indicate. this indicator is a bad idea because missingness will likely be correlated with your outcome and you'll think your model is great when you've just given it the answer to start with. see Harrell Regression Modelling Strategies for details.}
}
\value{
data frame with mean imputed data and possibly missingness indicators
}
\description{
For fast prototyping only!
}
