% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_hpd_int}
\alias{mcmc_hpd_int}
\title{Plot 95 percent HPD credible intervals for posterior samples}
\usage{
mcmc_hpd_int(postr_stats)
}
\arguments{
\item{postr_stats}{Data frame returned from \code{\link{mcmc_stats}}}
}
\value{
ggplot object
}
\description{
If the posterior sample data frame has a column \code{mle}, the column is
assumed to MLE point estimates of parameters and these are plotted next to
the credible intervals as red triangles.
}
