% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_stats}
\alias{mcmc_stats}
\title{Compute common statistics on posterior samples}
\usage{
mcmc_stats(df_mcmc)
}
\arguments{
\item{df_mcmc}{A data frame with posterior samples in a list column called
\code{samples}}
}
\description{
Compute common statistics on posterior samples
}
