% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_trace}
\alias{mcmc_trace}
\title{Create a trace plot from a dataframe of posterior samples}
\usage{
mcmc_trace(postr)
}
\arguments{
\item{postr}{Data frame of posterior samples}
}
\value{
ggplot object
}
\description{
Create a trace plot from a dataframe of posterior samples
}
