% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{mutate_by}
\alias{mutate_by}
\title{Mutate by group with less boilerplate}
\usage{
mutate_by(df, group_var, ...)
}
\arguments{
\item{df}{data frame}

\item{group_var}{bare/unquoted column name to group by}

\item{...}{expressions as you would normally use in mutate}
}
\value{
ungrouped data frame containing results of grouped summary
}
\description{
Mutate by group with less boilerplate
}
