% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{num_levels}
\alias{num_levels}
\title{Count the number of unique factor level combinations present in a data frame}
\usage{
num_levels(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Bare/unquoted factors of interest}
}
\value{
Integer count of unique factor level combinations
}
\description{
Count the number of unique factor level combinations present in a data frame
}
\examples{
num_levels(iris, Species)
}
