% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{pdf_table}
\alias{pdf_table}
\title{Add a nicely formatted table to a PDF knitr from R markdown}
\usage{
pdf_table(df, title = "TODO: ADD TITLE")
}
\arguments{
\item{df}{Data frame to print as a nice table}

\item{title}{Title of table}
}
\value{
Kable something?
}
\description{
Add a nicely formatted table to a PDF knitr from R markdown
}
