% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{perc}
\alias{perc}
\title{Generic to calculate percentages in the same vein as `dplyr::count`}
\usage{
perc(x, ...)
}
\arguments{
\item{x}{Object to summarize percentages for}

\item{...}{Additional arguments}
}
\value{
Percentage summary of \code{x}
}
\description{
Generic to calculate percentages in the same vein as `dplyr::count`
}
