% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{quad_form}
\alias{quad_form}
\title{Calculate the value of a quadratic form}
\usage{
quad_form(A, x)
}
\arguments{
\item{A}{Square n-by-n matrix}

\item{x}{Vector of length n}
}
\value{
Double value of the quadratic form x'Ax
}
\description{
Calculate the value of a quadratic form
}
