% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{rand_n_groups}
\alias{rand_n_groups}
\title{Select random groups from a dataframe with NSE}
\usage{
rand_n_groups(df, group_var, n)
}
\arguments{
\item{df}{A data frame or data frame like object recognized by \code{dplyr}}

\item{group_var}{Feature to group by. Uses NSE like dplyr verbs.}

\item{n}{Number of groups to take. If there are fewer than \code{n} groups, uses the}
}
\value{
Dataframe with n random groups selected according to specified grouping feature.
}
\description{
Groups a dataframe by the specified feature, selects n of these groups and
discards the rest. Returned dataframe is ungrouped.
}
\examples{
rand_n_groups(mtcars, cyl, 2)

}
