% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{remove_na}
\alias{remove_na}
\title{Remove all NA values for a vector}
\usage{
remove_na(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector with all NA values removed
}
\description{
Remove all NA values for a vector
}
\examples{
remove_na(c(2, NA, 3, NaN))
}
