% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rmse}
\alias{rmse}
\title{Get the RMSE of a set of predictions}
\usage{
rmse(predicted, true)
}
\arguments{
\item{predicted}{Predicted values}

\item{true}{True values}
}
\value{
RMSE between true and predicted values
}
\description{
Get the RMSE of a set of predictions
}
