% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{rmvn}
\alias{rmvn}
\title{Sample from the multivariate normal distribution}
\usage{
rmvn(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{mu}{Center of the distribution.}

\item{sigma}{Covariance matrix.}
}
\value{
A matrix where each row is a sample.
}
\description{
Sample from the multivariate normal distribution
}
\examples{

rmvn(5, 1:3, diag(3))
}
