% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvn_arma}
\alias{rmvn_arma}
\title{Sample from the multivariate normal distribution}
\usage{
rmvn_arma(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{mu}{Center of the distribution.}

\item{sigma}{Covariance matrix.}
}
\value{
A matrix where each row is a sample.
}
\description{
Sample from the multivariate normal distribution
}
\examples{

mu <- 1:3
sigma <- diag(3)
rmvn(10, mu, sigma)

}
