% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{rmvn_ql}
\alias{rmvn_ql}
\title{Sample efficiently from the MVN parameterized by Q matrix and ell vector}
\usage{
rmvn_ql(Q, ell)
}
\arguments{
\item{Q}{Precision matrix for MVN distribution}

\item{ell}{Multiplicative factor combined with \code{Q} to get mean of MVN dist}
}
\description{
i.e draw a single sample from \deqn{MVN(\ell Q^{-1}, Q^{-1})}
}
