% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{sample_n_feats}
\alias{sample_n_feats}
\title{Sample n features from a data frame randomly}
\usage{
sample_n_feats(df, n)
}
\arguments{
\item{df}{Data frame to sample features from}

\item{n}{Number of columns to randomly select}
}
\value{
Sample subsample of features in a dataframe
}
\description{
Sample n features from a data frame randomly
}
\examples{
sample_n_feats(mtcars, 3)
}
