% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{spread_keep}
\alias{spread_keep}
\title{Spread, but keep the key column}
\usage{
spread_keep(df, key, value, ...)
}
\arguments{
\item{df}{data frame}

\item{key}{bare/unquoted key as in \code{spread}}

\item{value}{bare/unquoted value as in \code{spread}}

\item{...}{additional arguments to \code{spread}}
}
\value{
spread data frame with key column retained
}
\description{
Spread, but keep the key column
}
