% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{spread_keep_}
\alias{spread_keep_}
\title{Spread, but keeps the key column and uses standard evaluation}
\usage{
spread_keep_(df, key, value, ...)
}
\arguments{
\item{df}{data frame}

\item{key}{key as in \code{spread}, string}

\item{value}{value as in \code{spread}, string}

\item{...}{additional arguments to \code{spread}}
}
\value{
spread data frame with key column retained
}
\description{
Spread, but keeps the key column and uses standard evaluation
}
