% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{spread_repeated}
\alias{spread_repeated}
\title{Repeatedly spread on a single key column}
\usage{
spread_repeated(df, key, ..., fill = NA, convert = FALSE, drop = TRUE,
  sep = NULL)
}
\arguments{
\item{df}{data frame}

\item{key}{bare/unquoted key column as in \code{spread}}

\item{...}{tidyselect specification of columns to spread on}

\item{fill}{as in \code{spread}}

\item{convert}{as in \code{spread}}

\item{drop}{as in \code{spread}}

\item{sep}{as in \code{spread}}
}
\value{
data frame repeatedly spread on key column
}
\description{
dplyr spread generalization to spread multiple columns from the same key
uses select semantics for convenience, and gives the produced columns
informative names using the add_detail function
}
