% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{summarize_by}
\alias{summarize_by}
\title{Summarize by group with less boilerplate}
\usage{
summarize_by(df, group_var, ...)
}
\arguments{
\item{df}{data frame}

\item{group_var}{bare/unquoted column name to group by}

\item{...}{expressions as you would normally use in summarize}
}
\value{
ungrouped data frame containing results of grouped summary
}
\description{
Summarize by group with less boilerplate
}
