% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{t_df}
\alias{t_df}
\title{Transpose a dataframe}
\usage{
t_df(df)
}
\arguments{
\item{df}{An unfortunately transposed data frame}
}
\value{
Transposed dataframe with appropriate column types
}
\description{
Transpose a dataframe without accidentally converting everything to
character. Assumes first column of input is a vector of feature names.
If any of the data is non-numeric you're SOL and tranposed cols will
be character as well.
}
