% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{t_mcmc}
\alias{t_mcmc}
\title{Collect posterior samples into a list column}
\usage{
t_mcmc(df)
}
\arguments{
\item{df}{Data frame of posterior samples where each row corresponds to a sample}
}
\value{
Data frame with two columns: one containing parameter names, one containing
  a vector of parameter samples
}
\description{
Collect posterior samples into a list column
}
