% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{top_n_groups}
\alias{top_n_groups}
\title{Select rows in the top n groups according to some summarization operation}
\usage{
top_n_groups(df, group_var, val_var, val_fun = sum)
}
\arguments{
\item{df}{data frame}

\item{group_var}{grouping variable}

\item{val_var}{value assessment variable to be summarized by group}

\item{val_fun}{value assessment function to apply to \code{val_var} by group}
}
\value{
data frame only with observations from top groups
}
\description{
Select rows in the top n groups according to some summarization operation
}
