% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{var_plot}
\alias{var_plot}
\title{Plot the variable importance of a ranger object}
\usage{
var_plot(rf)
}
\arguments{
\item{rf}{Object of class "ranger"}
}
\value{
\code{ggplot2} bar plot of variable importance
}
\description{
Plot the variable importance of a ranger object
}
