% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayeslib-package.R
\docType{package}
\name{hayeslib}
\alias{hayeslib}
\alias{hayeslib-package}
\title{hayeslib: a package to learn about packages. #meta}
\description{
Once the data manipulation functions get cleaned up there'll be a
little explanation of what's available here.
}
